requirements.txt
alpaca-py>=0.21.0
pandas>=2.0.0
pytz>=2024.1
python-dotenv>=1.0.0
